connected search events, login modal, and registration to current template. Added some photos intended for the carousel, still need to get them working
