+++ controllers/static_controller.php
@@ -11,6 +11,11 @@ class StaticController{
     public function registration() {
         require_once('views/static/registration.php');
     }
+
+    public function searchEvents(){
+        require_once('views/static/search_events.php');
+    }
+
 }
 
 ?>
