+++ views/static/signUp.php
@@ -1,0 +1,75 @@
+<!--<!doctype html>-->
+<!---->
+<!--<html lang="en">-->
+<!--<head>-->
+<!--    <meta charset="utf-8">-->
+<!---->
+<!--    <title>Sign Up Page</title>-->
+<!--    <meta name="description" content="Sign Up Page">-->
+<!--    <meta name="author" content="SitePoint">-->
+
+<!--    <link rel="stylesheet" href="views/styles/styles.css"/>-->
+    <!--[if lt IE 9 compatibility]>
+    <!--<script src="https://cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.js"></script>-->
+<!--    <![endif]-->
+<!--</head>-->
+
+<!--<body>-->
+<!--<h2>Modal Login Form</h2>-->
+<!-- Dispose of the button onclick entry below when integrating signUp page into project -->
+<!--<button onclick="document.getElementById('id01').style.display='block'" style="width:auto;">Login</button>-->
+
+<!--<div id="id01" class="modal">-->
+    <!--action_page.php is the php file which will perform user validation when the 'submit' button is clicked -->
+<!--    <form class="modal-content animate" action="/action_page.php">-->
+<!--        <div class="img-container">-->
+<!--            <span onclick="document.getElementById('id01').style.display='none'" class="close" title="Close Modal">&times;</span>-->
+<!--            <div><img src="../images/Fun_Logo1.png" alt="Logo" class="logo"></div>-->
+<!--            <img src="../images/img_avatar2.png" alt="Avatar" class="avatar">-->
+<!---->
+<!--            <div>-->
+                <!-- Add icon library from font-awesome -->
+<!--                <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">-->
+
+                <!-- Add font awesome icons -->
+<!--                <a href="#" class="fa fa-facebook"></a>-->
+<!--                <a href="#" class="fa fa-twitter"></a>-->
+<!--            </div>-->
+<!--        </div>-->
+<!---->
+<!--        <div class="container">-->
+<!--            <label for="uname"><b>Username</b></label>-->
+<!--            <input type="text" placeholder="Enter Username" name="user-name" required>-->
+<!---->
+<!--            <label for="psw"><b>Password</b></label>-->
+<!--            <input type="password" placeholder="Enter Password" name="psw" required>-->
+<!---->
+<!--            <button type="submit">Login</button>-->
+<!--            <label>-->
+<!--                <input type="checkbox" checked="checked" name="remember"> Remember me-->
+<!--            </label>-->
+<!--        </div>-->
+<!---->
+<!--        <div class="container" style="background-color:#FAFC2F">-->
+<!--            <button type="button" onclick="document.getElementById('id01').style.display='none'" class="cancel-btn">Cancel</button>-->
+<!--            <span class="psw">Forgot <a href="#">password?</a></span>-->
+<!---->
+<!--        </div>-->
+<!--    </form>-->
+<!--</div>-->
+
+<!--<script>-->
+<!--    // Get the modal-->
+<!--    var modal = document.getElementById('id01');-->
+<!---->
+<!--    // When the user clicks anywhere outside of the modal, the modal with close-->
+<!--    window.onclick = function(event) {-->
+<!--        if (event.target == modal) {-->
+<!--            modal.style.display = "none";-->
+<!--        }-->
+<!--    }-->
+<!--</script>-->
+
+
+<!--</body>-->
+<!--</html>-->
